+++ chat/src/__init__.py
@@ -1,0 +1,1 @@
+"""chat src package initializer."""